+++ ignite/chainconfig/errors.go
@@ -15,7 +15,7 @@ type ValidationError struct {
 	Message string
 }
 
-func (e *ValidationError) Error() string {
+func (e ValidationError) Error() string {
 	return fmt.Sprintf("config is not valid: %s", e.Message)
 }
 
@@ -24,6 +24,15 @@ type UnsupportedVersionError struct {
 	Version config.Version
 }
 
-func (e *UnsupportedVersionError) Error() string {
-	return fmt.Sprintf("config version %d is not supported", e.Version)
+func (e UnsupportedVersionError) Error() string {
+	return fmt.Sprintf("config version %s is not supported", e.Version)
+}
+
+// VersionError is returned when config version is not the latest.
+type VersionError struct {
+	Version config.Version
+}
+
+func (e VersionError) Error() string {
+	return fmt.Sprintf("config version %s is required and the current version is %s", LatestVersion, e.Version)
 }